Design update profile page
If the session is not active then redirect to index page.

The update form shows the current user, firstname and lastname. To
submit the data is needed enter the current password.

The form submit redirects to updateprocess, this verify the entered
account password, also update the corresponding cell for the user if the
column value has changed.

If the update results in success it redirects to index page, it doesn't
then shows the correspondig message.